+++ spec/telegram/bot/middleware_spec.rb
@@ -5,7 +5,8 @@ require 'rack/mock'
 RSpec.describe Telegram::Bot::Middleware do
   let(:instance) { described_class.new bot, controller }
   let(:bot) { double(:bot) }
-  let(:controller) { double(:controller, dispatch: :dispatch_result) }
+  let(:controller) { double(:controller, dispatch: nil) }
+  let(:webhook_response) { nil }
 
   describe '#call' do
     subject { instance.call(env) }
@@ -25,6 +26,15 @@ RSpec.describe Telegram::Bot::Middleware do
       let(:instance) { ActionDispatch::ParamsParser.new(super()) }
     end
 
+    before do
+      # `dispatch` claims the webhook response, when there's one, by setting it
+      # on the very same request object it received - same as the real controller.
+      allow(controller).to receive(:dispatch) do |_bot, _update, request|
+        key = Telegram::Bot::UpdatesController::WEBHOOK_RESPONSE_ENV_KEY
+        request.set_header(key, webhook_response)
+      end
+    end
+
     it 'calls dispatch on controller' do
       expect(controller).to receive(:dispatch).
         with(bot, update, instance_of(ActionDispatch::Request))
@@ -32,5 +42,10 @@ RSpec.describe Telegram::Bot::Middleware do
     end
 
     it { should eq [200, {}, ['']] }
+
+    context 'when controller claims the webhook response' do
+      let(:webhook_response) { '{"method":"sendMessage","text":"hi"}' }
+      it { should eq [200, {'Content-Type' => 'application/json'}, [webhook_response]] }
+    end
   end
 end